modified:   src/headers/app.helper.h 	modified:   src/helper.c 	modified:   src/my_printf.c

+++ src/headers/app.helper.h
@@ -1,7 +1,7 @@
 #ifndef MY_PRINTF_APP_HELPER_H
 #define MY_PRINTF_APP_HELPER_H
 
-extern size_t   stringLengthHelper(string s);
+extern long int stringLengthHelper(string s);
 extern void     charPrintHelper(const char c);
 extern void     stringPrintHelper(const string s);
 extern void     integerPrintHelper(int n);
